feat(fe:FSADT1-539) - refactor the update form value functions - update the way to display error message, create a error property and computed based on the validation result, compute fieldprops based on the error - fix add row problem for table/group refs: #255
