+++ frontend/src/common/FormComponentOptions.vue
@@ -2,14 +2,14 @@
   <div>
     <FormInput
       v-if="schema.type == 'input'"
-      :fieldProps="schema.fieldProps"
+      :fieldProps="computedFieldPropsForSingleComponent(schema.fieldProps)"
       :value="data"
       :disabled="schema.disabled ? true : false"
       @updateValue="updateFormValue"
     />
     <FormSelect
       v-if="schema.type == 'select'"
-      :fieldProps="schema.fieldProps"
+      :fieldProps="computedFieldPropsForSingleComponent(schema.fieldProps)"
       :value="data"
       :options="schema.options"
       :disabled="schema.disabled ? true : false"
@@ -17,14 +17,14 @@
     />
     <FormCheckbox
       v-if="schema.type == 'checkbox'"
-      :fieldProps="schema.fieldProps"
+      :fieldProps="computedFieldPropsForSingleComponent(schema.fieldProps)"
       :value="data"
       :disabled="schema.disabled ? true : false"
       @updateValue="updateFormValue"
     />
     <FormCheckboxGroup
       v-if="schema.type == 'checkboxgroup'"
-      :fieldProps="schema.fieldProps"
+      :fieldProps="computedFieldPropsForSingleComponent(schema.fieldProps)"
       :value="data"
       :options="schema.options"
       :disabled="schema.disabled ? true : false"
@@ -32,7 +32,7 @@
     />
     <FormRadioGroup
       v-if="schema.type == 'radiogroup'"
-      :fieldProps="schema.fieldProps"
+      :fieldProps="computedFieldPropsForSingleComponent(schema.fieldProps)"
       :value="data"
       :options="schema.options"
       :disabled="schema.disabled ? true : false"
@@ -44,6 +44,7 @@
       :addButtonText="schema.addButtonText"
       :deleteButtonText="schema.deleteButtonText"
       :subfields="schema.subfields"
+      :error="error"
       :fieldProps="schema.fieldProps"
       @updateFormArrayValue="updateFormArrayValue"
       @addRow="addRow"
@@ -54,6 +55,7 @@
       :data="data"
       :addButtonText="schema.addButtonText"
       :subfields="schema.subfields"
+      :error="error"
       :fieldProps="schema.fieldProps"
       @updateFormArrayValue="updateFormArrayValue"
       @addRow="addRow"
@@ -63,6 +65,7 @@
 </template>
 
 <script setup lang="ts">
+import { computed } from "vue";
 import type { PropType } from "vue";
 import FormInput from "./FormInput.vue";
 import FormSelect from "./FormSelect.vue";
@@ -71,7 +74,11 @@ import FormCheckboxGroup from "./FormCheckboxGroup.vue";
 import FormRadioGroup from "./FormRadioGroup.vue";
 import FormGroup from "./FormGroup.vue";
 import FormTable from "./FormTable.vue";
-import type { FormComponentSchemaType } from "../core/FormType";
+import type {
+  FormComponentSchemaType,
+  FormFieldValidationResultType,
+  FormFieldTemplateType,
+} from "../core/FormType";
 
 const props = defineProps({
   schema: {
@@ -82,6 +89,19 @@ const props = defineProps({
     type: [String, Object, Number, Array, Boolean],
     required: true,
   },
+  error: {
+    type: Array<FormFieldValidationResultType>,
+    default: [],
+  },
+});
+
+const computedFieldPropsForSingleComponent = computed(() => {
+  return (fieldProps: FormFieldTemplateType) => {
+    if (props.error.length > 0) {
+      return { ...fieldProps, errorMsg: props.error[0].errorMsg };
+    }
+    return fieldProps;
+  };
 });
 
 const emit = defineEmits([
