+++ frontend/src/common/FormTable.vue
@@ -13,6 +13,7 @@
             <FormComponentOptions
               :data="row[column.fieldProps.id]"
               :schema="column"
+              :error="computedError(column.fieldProps.id, rowIndex)"
               @updateFormValue="
                 (id, newValue) => updateFormArrayValue(id, newValue, rowIndex)
               "
@@ -38,7 +39,7 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from "vue";
+import { computed } from "vue";
 import type { PropType } from "vue";
 import FormFieldTemplate from "./FormFieldTemplate.vue";
 import FormComponentOptions from "./FormComponentOptions.vue";
@@ -48,6 +49,7 @@ import type {
   FormFieldTemplateType,
   FormComponentSchemaType,
   CommonObjectType,
+  FormFieldValidationResultType,
 } from "../core/FormType";
 
 const props = defineProps({
@@ -61,6 +63,17 @@ const props = defineProps({
     required: true,
   },
   fieldProps: Object as PropType<FormFieldTemplateType>,
+  error: {
+    type: Array<FormFieldValidationResultType>,
+    default: [],
+  },
+});
+
+const computedError = computed(() => {
+  return (subFieldId: string, rowIndex: number) =>
+    props.error.filter(
+      (each) => each.subFieldId == subFieldId && each.rowIndex == rowIndex
+    );
 });
 
 const emit = defineEmits(["updateFormArrayValue", "addRow", "deleteRow"]);
