+++ frontend/src/common/examples/FormJsonSchemaExample.vue
@@ -27,12 +27,6 @@
 import { ref } from "vue";
 import FormComponentOptions from "../FormComponentOptions.vue";
 import CollapseCard from "../CollapseCard.vue";
-import type {
-  FormFieldTemplateType,
-  FromSelectOptionType,
-  FormCheckBoxGroupOptionType,
-  FormRadioGroupOptionType,
-} from "../../core/FormType";
 
 /* ------- container props ----------*/
 const containerProps = {
