+++ frontend/src/pages/applynewclient/formsections/FormSections.vue
@@ -3,102 +3,26 @@
     <!------ begin section ------->
     <FormSectionTemplate
       :data="formData.state.begin"
-      :sectionProps="computedBeginSchema"
-      @updateFormValue="
-        (fieldId, newValue) =>
-          formData.mutations.updateFormValue('begin', fieldId, newValue)
-      "
-      @updateFormArrayValue="
-        (fieldId, subFieldId, newValue, rowIndex) =>
-          formData.mutations.updateFormArrayValue(
-            'begin',
-            fieldId,
-            subFieldId,
-            newValue,
-            rowIndex
-          )
-      "
-      @addRow="(fieldId) => formData.mutations.addRow('begin', fieldId)"
-      @deleteRow="
-        (fieldId, rowIndex) =>
-          formData.mutations.deleteRow('begin', fieldId, rowIndex)
-      "
+      :sectionProps="beginSectionSchema"
     />
 
     <!------ company/individual information section ------->
     <FormSectionTemplate
       v-if="computedInformationSchemaType !== ''"
       :data="formData.state.information"
-      :sectionProps="computedInformationSchema[computedInformationSchemaType]"
-      @updateFormValue="
-        (fieldId, newValue) =>
-          formData.mutations.updateFormValue('information', fieldId, newValue)
-      "
-      @updateFormArrayValue="
-        (fieldId, subFieldId, newValue, rowIndex) =>
-          formData.mutations.updateFormArrayValue(
-            'information',
-            fieldId,
-            subFieldId,
-            newValue,
-            rowIndex
-          )
-      "
-      @addRow="(fieldId) => formData.mutations.addRow('information', fieldId)"
-      @deleteRow="
-        (fieldId, rowIndex) =>
-          formData.mutations.deleteRow('information', fieldId, rowIndex)
-      "
+      :sectionProps="informationSectionSchema[computedInformationSchemaType]"
     />
 
     <!------ contact information section ------->
     <FormSectionTemplate
       :data="formData.state.contact"
-      :sectionProps="computedContactSectionSchema"
-      @updateFormValue="
-        (fieldId, newValue) =>
-          formData.mutations.updateFormValue('contact', fieldId, newValue)
-      "
-      @updateFormArrayValue="
-        (fieldId, subFieldId, newValue, rowIndex) =>
-          formData.mutations.updateFormArrayValue(
-            'contact',
-            fieldId,
-            subFieldId,
-            newValue,
-            rowIndex
-          )
-      "
-      @addRow="(fieldId) => formData.mutations.addRow('contact', fieldId)"
-      @deleteRow="
-        (fieldId, rowIndex) =>
-          formData.mutations.deleteRow('contact', fieldId, rowIndex)
-      "
+      :sectionProps="contactSectionSchema"
     />
 
     <!------ add authorized individual section ------->
     <FormSectionTemplate
       :data="formData.state.authorized"
-      :sectionProps="computedAuthorizedSectionSchema"
-      @updateFormValue="
-        (fieldId, newValue) =>
-          formData.mutations.updateFormValue('authorized', fieldId, newValue)
-      "
-      @updateFormArrayValue="
-        (fieldId, subFieldId, newValue, rowIndex) =>
-          formData.mutations.updateFormArrayValue(
-            'authorized',
-            fieldId,
-            subFieldId,
-            newValue,
-            rowIndex
-          )
-      "
-      @addRow="(fieldId) => formData.mutations.addRow('authorized', fieldId)"
-      @deleteRow="
-        (fieldId, rowIndex) =>
-          formData.mutations.deleteRow('authorized', fieldId, rowIndex)
-      "
+      :sectionProps="authorizedSectionSchema"
     />
   </div>
 </template>
@@ -111,22 +35,7 @@ import { beginSectionSchema } from "../formsectionschemas/BeginSectionSchema";
 import { informationSectionSchema } from "../formsectionschemas/InformationSectionSchema";
 import { contactSectionSchema } from "../formsectionschemas/ContactSectionSchema";
 import { authorizedSectionSchema } from "../formsectionschemas/AuthorizedSectionSchema";
-import { formData } from "../../../helpers/FormState";
-import { addErrMsgToSchema } from "../../../helpers/formvalidation/AddErrorToSchema";
-
-// create computed section schemas to determine when to display error messages as specified in the validation result
-const computedBeginSchema = computed(() => {
-  return addErrMsgToSchema(beginSectionSchema, "begin");
-});
-const computedInformationSchema = computed(() => {
-  return addErrMsgToSchema(informationSectionSchema, "information");
-});
-const computedContactSectionSchema = computed(() => {
-  return addErrMsgToSchema(contactSectionSchema, "contact");
-});
-const computedAuthorizedSectionSchema = computed(() => {
-  return addErrMsgToSchema(authorizedSectionSchema, "authorized");
-});
+import { formData } from "../../../store/newclientform/FormData";
 
 // based on client type, show different schema contenct for the information section
 const computedInformationSchemaType = computed(() => {
